Include tentatively pi.proto in the PI build (#34)
I added googleapis (https://github.com/googleapis/googleapis) as a
submodule so that the required protos are available. When running the
configure script with --with-proto, we will check if all dependencies
are satisfied (protoc, libproto, libgrpc) and tell make that it can
descend into the proto/ directory when building. In the proto/ directory
we run protoc and compile a dummy library with the generated sources.