Add modify_wkey and delete_wkey (#40)
These functions lets you modify or delete a table entry by re-using the
same match key that you provided when adding the entry (instead of the
returned entry handle). We consider that the priority is part of the
match key. Therefore, the priority has been moved out of the table entry
properties, to the match key.