Define DIST_SUBDIRS explicitly when needed
We run AC_CONFIG_SUBDIR conditionally for the proto directory (to avoid
having to install dependencies when protobug + gRPC not needed). Which
means that proto/Makefile is conditionally generated. When running make
dist, this is an issue because DIST_SUBDIRS is set by automake to
contain all suddirs, including proto. We therefore set DIST_SUBDIRS
manually when needed to avoid errors.