Add function call counter to dummy target code
Console logging is now conditional on the DEBUG preprocessor flag. Each
function call increments a counter, with each function having its own
counter. A Judy array is used for the couters, to map function names to
the counter values. This is convenient to check the function was called
the appropriate number of time when doing some performance testing.