Updated demo_grpc to new pi.proto
The demo is now part of the repo's automake build. I am not sure it is a
good solution for the long term, but in the mean-time it is convenient
(i.e. before we clean-up which libraries are installed, under which
name). Still things to cleanup in the demo, but can be done later.