Proposal for proto frontend unit tests
We propose a framework built with googletest and googlemock. We are
testing the following path: PI proto frontend -> PI C++ frontend -> PI
common code. If needed we can consider adding the prototype gRPC PI
server in this path to exercise it.

Right now we put a lot of stuff (DummySwitch code...) in a single file;
we will probably want to organize the classes and test cases better in
the future.