Much progress for action profile support
Action profiles are now supported by DeviceMgr. A lot of work was needed
because mapping the proto ids to PI-provided handles requires
maintaining some state in the DeviceMgr. Indirect ids (members or
groups) are now supported by the table entry add operation. A bunch of
unit tests were also added. Support for action profiles in the prototype
gRPC server (for the demo) will come in a subsequent commit.