Enable P4 programmers to choose PI ids (#68)
P4 programmers can now choose (part of) the PI id using the id pragma /
annotation. This has the advantage of ensuring that programmer-chosen
ids don't change when the P4 program is modified.

The P4 programmer can specify ids for objects which are represented in
the PI: tables, actions, counters, meters, action profiles (and to some
extent field lists). For fields and action parameters this is a little
trickier. For all cases the programmer can choose an id in the range [0,
0xffff]. For most resources, we just take this id and set the first byte
of the unit32 to the resource type (the second byte is therefore always
0). For header fields, the programmer can choose an id for the header
instance. The PI will then use the following id for fields: PI_FIELD_ID
(1 byte) | programmer-provided header id (2 bytes) | field index in
header type (1 byte). The same goes for action parameters. This is all a
little hacky and we hope P4_16 lets us do something cleaner.

For header stack instances, things are even harder. The P4 programmer
can provide a list of ids in the pragma instead of a single id. These
ids will be used for each header instance in the stack (or rather for
their fields).

To support arbitrary ids, we had to replace the data structure used to
store p4info information. It was no longer possible to use a simple
array / vector (previous ids were always allocated contiguously),
instead we now need to use a map, where the key is the id.

This PR also includes style fixes (I have updated from clang-format-3.6
to clang-format-3.8 which apparently reports more errors).