Fix shared action profile support
Shared action profiles were instantiated multiple times (once per table
referencing it), this has been fixed. Note that the bmv2 JSON reader now
requires the latest JSON format (>=2.0) and will throw an error if an
old JSON is used. A version check was added to the reader code to this
effect.

Pragmas are now correctly handled for action profiles. In particular an
ID can be forced for an action profile (shared or not) using the 'id'
pragma.

This commit also updates some older test JSON files.

This commit fixes issue #85