Rename PI gRPC service to P4Runtime
We only renamed the service and updated the demo code accordingly. The
proto file was also renamed from pi.proto to p4runtime.proto.

For now, the rest of the code still refers to PI. We may also want to
rename the repository from PI to p4Runtime / p4-runtime in the future,
but for now it seems that the gRPC service was the priority.