adds skeleton to enhance documentation (#96)
* adds skeleton to enhance documentation

Adds configuration to support doxygen documentation generation to
autotools as well as an initial skeleton to integrate documentation
that is not part of the source code.

For now, I kept it simple so that we can just add .md files as
"chapters" for different aspects of the documentation. Add a chapter
by creating a new file, prefixed by the chapter number (e.g. Chapter 1
Overview is 01_overview.md). The chapter number prefix is used instead
of creating a hierarchy of doxygen pages because those introduce extra
links that are a pain. However, doxygen consumes markdown very nicely,
and the only need is to place the files in the appropriate order.

If we want to be fancy with the documents generated (PDF and HTML)
then we need more work on the style sheets and we'll likely required
to generate custom header and footer files -- this will be a task for a
technical writer.

* addresses reviewer comments

Remove obsolete aminclude.m4
Add to the source distribution all the files required to build docs.