Add PI support for fetching action profile entries
Added pi_act_prof_entries_fetch and supporting functions to PI, very
much like what is done for tables. The function was implemented for the
bmv2 and the rpc targets.

Added CLI commands to create a group, add a member to a group and dump
an action profile; this was mostly done in order to be able to easily
test pi_act_prof_entries_fetch.