Add missing clear_grp_h calls in bm implementation
The bmv2 PI implementation code modifies the group handles returned by
the bmv2 switch to mark them as group handles (with a special 1-byte
prefix) before returning them to the PI client. This means that when the
client gives the handle to us in another PI API call, we need to clear
the mark before passing it on to bmv2.