Stop using 0 for params and match fields ids (#127)
It has been decided to stop using 0 in ids for action parameters and
table match fields. This is because in protobuf 3, 0 is the default
value for integers and there is an ambiguity between an id that has been
set to 0 and an id that wasn't set. Additionally this is more
consistent with the way ids used to be allocated (0 was used exclusively
as an invalid id).