Took a stab at a GnmiMgr
Unlike DeviceMgr, GnmigMgr manages multiple devices because of the way the gnmi
proto was written. We need to update DeviceMgr to also manage multiple devices
for consistency. In the future we can also merge the 2 into one class, but for
now we keep them separate for the sake of simplicity.

Just like DeviceMgr, GnmiMgr uses the PIMPL idiom. We do not provide an
implementation of it though, it has to be provided by the user at link
time. This is also subject to change.

All this code is more a POC than anything else at this stage.