Make DeviceMgr responsible for locking TableInfoStore
We need to guarantee that the local table state is updated consistently
with lower-level driver operations. Otherwise, in the event of multiple
worker threads for P4Runtime, this could lead to complications. We
therefore make a table lock available to DeviceMgr. This means that
operations on the same table cannot be done in parallel, but this is a
reasonnable restriction which is likely to be shared by lower level
drivers too.