Tentative support for error_details in grpc::Status (#148)
Since version 1.3.0, gRPC supports returning error details along with
the grpc::Status. This commit modifies the gRPC server implementation
for P4Runtime to leverage this feature. DeviceMgr methods return a
google::rpc::Status object which is converted to a grpc::Status object
using the grpc::SetErrorDetails helper function which ships with
gRPC. On the client side, the grpc::ExtractErrorDetails method can be
used to retrieve the error details.

Note that the google::rpc::Status proto contains a `repeated
google.protobuf.Any` field which will be used to carry
P4Runtime-specific error codes for each batch request, in addition to
the generic gRPC error code.

Because of the new dependency on error details, the configure.ac file
was updated to require a version >= 1.3.0 for gRPC.