Temporary revert linking with grpc++_error_details
The library includes generated proto code for google.rpc.Status. We also
include this code in our libpiproto library which creates an issue when
starting any of our binaries. As a temporary solution we stop linking
with libgrpc++_error_details and instead make a copy of the source code
for SetErrorDetails.