Fix default action set in P4Runtime DeviceMgr (#174)
When the match key is empty, it means "set default action", and
DeviceMgr should not try to "validate" the match key, which would result
in an error (because of missing fields).
Added GTest unit test.
Many issues remain with default actions (e.g. direct resources support)
and default action get has not been tested yet, but it is now possible
to set the default action.