Fix build issue
The addition of a test in the previous commit revealed an issue when
building with clang in optimized mode. Executabled linked to both
lpiproto and lpiprotobuf would not be able to start because of overlap
in static initialization code and inlining of some gRPC code. I replaced
lpiproto with lpiprotogrpc which only includes the auto-generated code
for gRPC.