First stab at implementing controller arbitration (#199)
* First stab at implementing controller arbitration

We switched from an asynchronous server implementation to a synchronous
one for the P4Runtime bidi stream, as it makes the code easier to reason
about and there isn't a requirement to support a huge number of
controller connections per switch.

We started implementing controller arbitration, which allows the switch
to maintain multiple controller connections simultaneously (to a single
master controller and several slaves). As of now, some compromises have
been made to avoid a brutal transition. For example we support unset
election ids for backward-compatibility with existing Write requests.

* Accept all Write requests if no master

* Add unit tests for master arbitration

* Add some documentation for arbitration