Add example PTF test for P4 Runtime
We add some infrastructure code to easily run PTF tests written with P4
Runtime accross different targets. The objective is to have a testsuite
which can be used to evaluate conformance (and potentially to benchmark)
different P4 Runtime implementations.

As an example, we provide a first P4 program, l3_fwd_host.p4, along with
a test. We also provide instructions on how to run this PTF test on the
bmv2 simple_switch_grpc target.