Initial support for gNMI & OpenConfig
We only support gNMI subscriptions in ONCE mode and the
openconfig-interfaces YANG model.

We choose to use sysrepo as a YANG data store and to process XPaths, to
avoid having to implement our own solution. Using sysrepo may not be
appropriate for a production product, and this support is mostly meant
as a proof-of-concept at this stage.

See proto/README.md for more information and instructions on how to test
out gNMI support.