Remove bmv2 from Docker image ancestry
The bmv2 dependency is only needed for testing. It is therefore logical
to have 2 docker images: one without bmv2 and one with bmv2 for CI
testing. This means that the bmv2 Docker image can in turn inherit from
the PI Docker image and will be able to include the simple_switch_grpc
binary.