Infer XPath namespace qualifier from gNMI path
gNMI does not assume that the module name is included in the path, as
long as there is no ambiguity, i.e. no overlap in the implemented
schemas. However, sysrepo does require the module name as a namespace
qualifier for the first element in the path. To retrieve the module
name, we iterate over all the schemas explicitly installed (aka
implementeed) in sysrepo and look at the root node(s) for each one. We
then associate that root node name to the corresponding schema name,
which lets us do the conversion from gNMI path to sysrepo XPath.

In order to implement this, we need to be able to differentiate between
modules which are installed / implemented in sysrepo and modules which
are simply imported, for which there is no data tree. For example,
openconfig-interfaces dependends on ietf-interfaces. Both YANG modules
define a "interfaces" tree. If we install both, we will end up with both
data trees and we will not be able to infer the correct namespace
qualifier (in our case oepnconfig-interfaces) from the gNMI path using
the technique described above. We therefore updated the install_yangs.sh
script to only install the models we intend to support.

The update to install_yangs.sh requires the ability to provide multiple
search directories to sysrepoctl. This is only supported in the latest
sysrepo version (0.7.2) which is why we have updated the version
requirement in the README.

Finally, this commit also installs the openconfig-platform YANG model,
even though we are not doing anything with it yet in bmv2.