Got rid of GnmiMgr and cleaned-up server unit tests
The GnmiMgr PIMPL class, which was added a while ago, didn't prove
useful in implementing the gNMI service for sysrepo. As a result I am
removing this class for now.

This commit also includes some cleanup to the P4Runtime server
gtests. In particular, we now let the OS bind to any available port,
instead of harcoding a port number for each test process.