Tentative support for gNMI.Set + gNMI unit tests
This commit introduces support for setting leaf data nodes with
gNMI.Set. We use sysrepo's sr_set_item_str to easily convert from
protobuf leaf types to sysrepo leaf types. We only support update and
delete for now, not replace operations (see gNMI specification for
difference in semantics between update and replace).

We also introduce a first gtest for the sysrepo gNMI
implementation. This is a test for gNMI.Set/update (leaf node). We will
be adding a test for gNMI.Set/delete and a test for gNMI.Subscribe/ONCE
very soon.

In order to run the gNMI tests in the Docker image, we need to make sure
that the sysrepo daemon is running, which is why we have defined an
ENTRY_POINT to start sysrepod in Dockerfile.bmv2 (which is the Docker
image we use for CI testing).

This commit also updates the openconfig/public submodule ref pointer (to
get the latest version of the Openconfig YANG models).