Require match field to be omitted for don't care
We have decided to require a match field to be omitted from the
TableEntry for don't care matches. This enables us to achieve smaller
protobuf messages while having a canonical representation for all
matches.

This commit implements the necessary changes to the P4Runtime server,
adds gtests and updates the P4 Runtime specification.

Fixes #35