Support reading from const tables with P4Runtime
It is possible that when using a constant table initializer in P4_16,
entries end up being added to the target out-of-band, i.e. without
P4Runtime. It is the case for example with bmv2, for which the entries
are added from the bmv2 JSON. We fixed the DeviceMgr implementation to
account for this case (and not assert if the entries returned by the
target are not present in the table store).

Fixes #295