Support reading direct resources with p4.TableEntry
We now support reading direct resources when reading a TableEntry with
P4Runtime. The values will only be included in the TableEntry message if
the corresponding fields are not unset.

In order to support this in P4Runtime, support also had to be added to
the PI library, and in particular pi_table_entries_fetch.