Support reading direct resources with p4.TableEntry (#321)
* Remove some un-needed indirection when reading table entry

* Support reading direct resources with p4.TableEntry

We now support reading direct resources when reading a TableEntry with
P4Runtime. The values will only be included in the TableEntry message if
the corresponding fields are not unset.

In order to support this in P4Runtime, support also had to be added to
the PI library, and in particular pi_table_entries_fetch.

* Cleanup direct resource handling in mock switch

Used for P4Runtime DevMgr testing.

* Add some basic match key filtering for table read in DevMgr

We still read all of the entries from the target and then do a
comparison on each one. The priority field is also used for the
comparison.