Support for arbitrary type serialization in P4Runtime
Adds support for arbitrary P4_16 types in P4Runtime, as discussed at the
03/21/2018 WG meeting (refer to minutes).

The changes incur minimal overhead for the most-common case (`bit<W>` in
P4, for which we use a binary string in P4Runtime), and enable us to
support complex types in lookup expressions and PSA externs such as
Register and Digest.

We include some documentation for the type specifications (P4Info) and
value representations (P4Runtime) which can be merged in the main
P4Info/P4Runtime specification in the future (before the v1.0 release).

To avoid breaking existing implementations, we do not use the new P4Data
representation for tables and value-sets.