Add support for P4Runtime RECONCILE_AND_COMMIT
As an action mode for SetForwardingPipelineConfig in reference
implementation.

We use a pretty naive implementation: we read all entries using a
wildcard read, perform the update sequence then write all entries back.

This commit is also the perfect opportunity to introduce proper
top-level locking to synchronize all RPCs. This introduces a dependency
on the Boost thread library as we want to use a shared mutex (not
available in the C++11 standard library). We allow concurrent Write
calls; however Read and SetForwardingPipelineConfig get an exclusive
lock. These Read / Write semantics are as agreed on for P4Runtime by the
P4 API WG.