Add definition of id prefixes to p4info.proto as enum
Historically ids have shared the same number-space (i.e. all ids for
top-level objects are unique). Furthermore, the compiler has been
allocating ids in such a way that it is possible to deduce the object
types (e.g. table, action, ...) based on the value of the
most-significant byte of the id.

We decided that it was a bad idea to change that scheme and break
existing implementations of P4Runtime. However, it was also decided that
we should add an enum to p4info.proto so that these 1-byte values are
documented in code somewhere. The compiler should use these values when
allocating ids for the P4 objects.