Add support for idle timeout for table entries in P4Runtime (#346)
* Add support for idle timeout for table entries in P4Runtime

Necessary changes to p4info.proto and p4runtime.proto to support idle
timeout for table entries. A table supports idle timeout if the
idle_timeout_behavior field in the corresponding P4Info Table message is
set to NOTIFY_CONTROL (we assume that this information comes from the P4
table definition, either as a table attribute or as an annotation).

When a table supports idle timeout, the P4Runtime client can specifiy a
TTL for each TableEntry. This TTL value will be returned to the client
during a Read operation. If an entry hasn't been hit in the last TTL
nanoseconds, a notification is sent to the client on the
StreamChannel. After a notification is sent, the P4Runtime server resets
the timer for this entry, which means a new notification will be sent
after another TTL nanoseconds if the entry is not hit. This means that
it is ok if notifications are "lost" or if the server drops
notifications because the channel / client is too busy.

When reading a TableEntry, the client can elect to read the elapsed time
(in nanoseconds) since the entry was last hit in the data-plane. We use
a protobuf message for this field in order to enable the client to
choose whether this value should be read and returned in the
response. This is consistent with what we do for direct resource fields.

We use a repeated field of TableEntry in the notification message so
that the server can coalesce notifications in one message when
appropriate (we need to define what appropriate means in the spec).

This may not be the most comprehensive proposal, but I believe it is
sufficient for P4Runtime v1.0 as it covers the general use case.

Fixes #265

* Added comment to IdleTimeoutNotification

To clarify which fields must be set in table_entry members.