Use p4v1 and p4configv1 namespace aliases
Instead of p4rt and p4config. This is based on a discussion with
reviewer @wmohsin about future-proofness, best practices and name
collisions.

Even though the fact that we were only using namespace aliases in
implementation files mean that there should be no name conflicts at link
time (namespace aliases are only visible in their translation unit and
do not affect the full-qualified name of linker symbols), we decided
that in the end it probably was simpler to just include the major
version number in the namespace alias. A change in major version number
should be a rare and disruptive event, and keeping the diff small when
updating the implementation may not be a valid concern. There will be
many things to figure out when & if we introduce P4Runtime v2 and it is
tough to predict how it will happen and which steps we will follow to
roll out the new version.