Remove VALID match type from P4Runtime
We remove the VALID match type from p4runtime.proto and
p4info.proto. The valid match type does not exist in P4_16, which means
that the p4c compiler never produces a VALID match type in P4Info even
for P4_14 programs, since all P4_14 programs are converted to P4_16
before P4Info is generated.

While the change is not backward-compatible since we are deleting
Protobuf fields, it should not impact existing Protobuf binaries that do
not use VALID, as it does not affect other Protobuf fields (we did not
change any of the Protobuf integer field values).

The VALID match type is preserved in the PI implementation for now, to
ensure backward-compatibility with the bmv2 JSON produced by the legacy
compiler (p4c-bm).

Fixes #367