Add implementation for P4Runtime MulticastGroupEntry message
This required defining some lower-level PI C functions to match the
implementation of other features. The PI functions are similar to the
ones exposed by bmv2.

This commit implements the new PI functions for bmv2 although the
implementation is not tested (unfortunately there is no simple way to
test without also implementing the functions for the deprecated rpc
target and in the CLI). However, we include unit tests for the mapping
from P4Runtime message to PI. Note that most people do not use this
implementation for bmv2 anyway, but instead use simple_switch_grpc.