Use NO_TIMEOUT as default value in P4Info
For the IdleTimeoutBehavior enum. In proto3, there is no need to use
UNSPECIFIED as the default value for enums (for backward-compatibility
purposes) as during deserialization, unrecognized enum values will be
preserved in the message. NO_TIMEOUT makes sense as the default and it
will make debug prints of the messages shorter (since most P4 tables
will use NO_TIMEOUT and default values are omitted from the debug
prints).

Fixes #379