Add support for serializable enums in P4Runtime
Proposal for supporting P4 serializable enums in P4Runtime. For each
serializable enum, P4Info includes the underlying value for each named
member of the enum. On the P4Runtime side, P4Data enables providing
either the name or the numerical value (it should be possible to provide
a value for a serializable enum even if this value does not correspond
to named member, as long as the numerical value falls into the range of
the underlying representation).

See P4_16 spec section 7.2.1 (Enumeration types) for more background on
safe enums vs. serializable enums.