Tentative support for Bazel build system
We add support for the Bazel build system. Support is probably not as
complete as for the original automake build system, and we recommend
users to stick to automake for now.

In particular the following can be used:
 * `bazel test //proto/tests:pi_proto_tests` will run the tests for the
   PI Protobuf frontend.
 * `bazel test //proto/tests:pi_proto_server_tests` will run the tests
   for the reference P4Runtime server implementation.
 * `bazel build //proto/server:pi_server` will build the PI Protobuf
   frontend library.

The following changes had to be made:
 * import libJudy as a Bazel repository (@judy).
 * use Absl/synchronization mutex instead of Boost mutex when building
   with Bazel since it seems hard to import Boost as a Bazel repository.
 * implement missing _pi_* functions for the GTest mock switch. With
   automake we tell the linker to ignore the missing symbols (which are
   not used at runtime) but Bazel seems to be using the Gold linker by
   default on Linux and we wanted a more robust solution (it turned out
   that only a few of the _pi_* target functions were not implemented
   anyway so there wasn't any good reason not to implement them; we will
   remove the automake hack in a future PR).
 * some #include directives had to be changed from quotes to system
   since Bazel uses `-iquote` for workspace include paths.

Finally a more general note on includes in Bazel. This project was not
written with Bazel in mind so we cannot follow all of Bazel "best
practices". Bazel assumes that all #include directives are relative to
the workspace root. The Bazel documentation advises against using the
`include` parameter to `cc_library` since it propagates to dependent
targets. However, when using `copts` instead we run into issues when
importing p4lang/PI as a dependency in another Bazel project - and we
don't want to have to set `copts` again in the that project. There may
be something we are missing here and there is probably room for
improvement. Maybe a future project :)