Fix default entry programming in P4Runtime implementation
Fix the implementation to match the P4Runtime specification, which
states that only the MODIFY update mode should be used for default
entries. We also include other fixes to make default entry support in
DeviceMgr more robust. There is still some work to do at the PI level
and DeviceMgr level to fully support default entries. At the moment, we
do not support reading the default entry and we do not support accessing
direct resources for the default entry (using DirectMeterEntry and
DirectCounterEntry).

Fixes #397