Support for one-shot action selector programming mode
P4Runtime has introduced a new mode to enable easier programming of
indirect tables which use an action selector. This new mode is called
"one shot". This commit introduces support for this in the reference
P4Runtime implementation built on top of PI.

Because one P4Runtime operation for one-shot programming may translate
to many PI function calls, we add the ability to do some rollback in
case of failure. This is not as powerful as the P4Runtime transation
semantics, but it should work for simple operations, like rolling back
creation of members & groups.

We chose to implement one-shot table MODIFY by first creating a new
group and adding the necessary members, then swapping the old group
pointer with the new one in the table entry. This means that we need
enough space in memory to hold the extra group. Note that the current
implementation always creates a group, even if the P4Runtime
ActionProfileActionSet only includes a single action.

The specification states that the 2 programming modes cannot be used
simultaneously for a given action selector. This commit implements this
check. If the action selector becomes empty, we enable the client to
switch modes.

Just like for "manual" action selector programming, we do not have any
support for weight and watch yet.

Finally this commit also adds the ASSERT_OK and EXPECT_OK macros for
gTest which enable us top check if the Status object returned by
DeviceMgr is OK in a less verbose fashion. Note that existing unit tests
haven't been updated yet to use these new macros.

Fixes #405