P4Runtime support for response_type in GetForwardingPipelineConfig
A recent change to P4Runtime adds a cookie to the
ForwardingPipelineConfig message and enables the client to choose what
to retrieve when performing a GetForwardingPipelineConfig RPC (cookie
only, or cookie + p4info, or cookie + device config). The cookie is
convenient for example to verify which config is currently on the
switch, without retrieving the entire device config.

This requires storing the device config in the server. I chose to store
it in a temporary file, rather than in memory, as it can be quite large.

This commit also updates the p4runtime submodule ref point to 1.0.0-rc3.