Add selector groups size checks as per P4Runtime specification
We enforce everything we can in the DeviceMgr implementation and return
the appropriate error code. If both the dynamic max size (provided at
group creation time) and the statis max size (from P4Info) are 0,
pi_act_prof_grp_create is called with a max_size of 0. It is the
responsibility of the PI target implementation to behave accordingly to
the spec then ("the target should use the maximum value it can
support").

Fixes #452