Support for weighted members in ActionProfileGroup
The support for weighted members assume that the underlying PI
implementation has no native support for weights.

 * for each member_id, we keep track of the maximum weight for that
   member (across all groups).
 * if the maximum weight is W, maintain W copies of the member (created
   by calling pi_act_prof_mbr_create W times with the same parameters)
   with W different handles.
 * if a group includes the member with weight w (where w is <= W), then
   pick w unique handles among the set of W handles we have and set the
   group membership using all w handles.
 * delete members that are no longer needed (when W decreases).

We also support weights for actions when using one-shot programming. In
this case we also create multiple copies to achieve weighted
selection. However, it is easier in this case because of the lack of
indirection (no sharing across groups).