Deprecate temporary P4DeviceConfig message
The assumption was always that this was a temporary message. It was
required when using bmv2 simple_switch, but is no longer needed with
simple_switch_grpc.

As a result of this change, the
ForwardingPipelineConfig::p4_device_config binary string is passed on
directly to PI. For backward compatibility, the DeviceMgr still tests
whether this field is a serialized P4DeviceConfig message (and behaves
as before).