Change locking model in DeviceMgr
Before processing each WriteRequest batch, we now ensure that the thread
processing the batch has exclusive access to all the objects modified in
the batch. This means that the underlying target can safely assume that
individual P4 forwarding elements are not accessed by concurrent
threads. This probably has some non neglectable overhead as we need to
iterate over the entire batch before we start processing it and we have
to perform some set operations.

This change means that we can probably get rid of some other mutexes in
the implementation (e.g. the ActionProfMgr mutex), but this needs to be
done carefully. We may be able to change TableStore locking as well; one
issue is the idle-timeout thread that accesses the store to convert the
PI match key to a P4Runtime match key.

Add new class of access in DeviceMgr