Fix concurrency issue in WatchPortEnforcer
WatchPortEnforcer::get_port_status was not thread-safe; we fix this
issue by having one port oper status map for each action profile
instance and by adding the action profile id as a parameter to
get_port_status.

Because there is no guarantee that all ports are configured before the
DeviceMgr instance is created, we add another PI function,
pi_port_status_get which can be used to synchronously query the status
of the port.