Enable WatchPortEnforcer to update action profiles in any order
When a port status event is generated. This ensures that WriteRequests
which modify a specific action profile do not affect processing of port
status events for other action profiles.

Also refactor AccessArbitration to use better RAII style. I realized
that I didn't delete copy & move operations for all the Access
objects. Based on current usage this was probably not an issue because
of the compiler using NRVO, but it could have led to nasty bugs in the
future.