Support for watch port attribute for ActionProfileGroup members (#471)
* Support for watch port attribute for ActionProfileGroup members

Watch ports are handled entirely by the DeviceMgr. The PI layer is
responsible for generating notifications (through callbacks) when the
status of a DP port changes. The WatchPortEnforcer instance (owned by
the DeviceMgr) implements a task queue which will process the
notifications asynchronously and activate / deactivate action profile
group members as needed.

The notion of activated / deactivated member is new. A deactivated
member is still present in a group (and will show as part of the group
when fetching the list of members) but cannot be selected by the data
path for packet forwarding any more.

We introduce 2 new PI APIs, pi_act_prof_grp_activate_mbr and
pi_act_prof_grp_deactivate_mbr, which targets are required to implement,
and which are used by the WatchPortEnforcer to activate / deactivate
members as needed when receiving port status event notifications from
the target.

When using the individual add / remove APIs to manage groups, we assume
that members are always added in the "activated" state. When using the
set members API to manage groups, we provide a new function parameter
which is used to choose whether each member should be activated /
deactivated.

I briefly toyed with the idea of letting the target implement support
for watch port itself, but ultimately decided against it. Some targets
may have optimized ways of dealing with ports going down however, and if
this is the case we may consider introducing the ability to delegate
watch support to the target in the future.

* Add unit tests for WatchPortEnforcer

* Fix AccessArbitration for pipeline config updates

There was a potential deadlock in the WatchPortEnforcer when the P4
changes. unique_access was renamed update_access. When requesting
"NoWriteAccess", one case use a new overload which "fails" if there is
an ongoing config update.

* Fix concurrency issue in WatchPortEnforcer

WatchPortEnforcer::get_port_status was not thread-safe; we fix this
issue by having one port oper status map for each action profile
instance and by adding the action profile id as a parameter to
get_port_status.

Because there is no guarantee that all ports are configured before the
DeviceMgr instance is created, we add another PI function,
pi_port_status_get which can be used to synchronously query the status
of the port.

* Enable WatchPortEnforcer to update action profiles in any order

When a port status event is generated. This ensures that WriteRequests
which modify a specific action profile do not affect processing of port
status events for other action profiles.

Also refactor AccessArbitration to use better RAII style. I realized
that I didn't delete copy & move operations for all the Access
objects. Based on current usage this was probably not an issue because
of the compiler using NRVO, but it could have led to nasty bugs in the
future.

* Add debug logging for activate / deactivate member

Unfortunately the logging messages currently use the PI internal handles
instead of the P4Runtime ids (for action profile member / group). We
should change that in the future.